moving demo to mex demo
demo was really more meant for the mex interface. You could use a similar interface for the command line or something but since you have the source I think it'll be easier to just include the files you need for the c demo
