+++ mexdemo/Makefile
@@ -1,6 +1,6 @@
 include ../config.mak
 
-TARGET_NAME = demo
+TARGET_NAME = mexdemo
 LDLIBS += -L$(BUILD_DIR) -lcsim2
 
 include ../shared_lib.mak
